bullet logic inmplemented and tested

+++ src/bullet.h
@@ -3,7 +3,6 @@
 #include "raylib.h"
 #include "math.h"
 
-
 class Bullet{
     public:
         Bullet(Vector2 startPos, Vector2 startSpeed, Texture2D* bulletTexture);
@@ -12,11 +11,12 @@ class Bullet{
         void Update();
         bool IsActive();
         Vector2 GetPosition();
+        
     private:
         Vector2 position;
         Vector2 speed;
         Texture2D* image;
-        int lifetime;
+        float lifetime;
         bool active;
